moving normalise to stark folder
