+++ src/lib/stark/formatStarkKey.js
